% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_readable_edge_label}
\alias{add_readable_edge_label}
\title{add_readable_edge_label}
\usage{
add_readable_edge_label(color = "label_colour", angle_calc = "along", size = 3)
}
\arguments{
\item{color}{color mapping name in the previous layer}

\item{angle_calc}{only `along` or `none`}

\item{size}{size of the label}
}
\value{
ggplot2 object
}
\description{
Add readable edge label by shadowtext.
Should be just under the geom_edge_* with the filtered labels.
}
\examples{
test_pathway <- create_test_pathway()
plt <- ggraph(test_pathway, layout="manual", x=x, y=y) +
 geom_edge_link(aes(label=subtype_name))+
 add_readable_edge_label()
}
