% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_functions.R
\name{add_title}
\alias{add_title}
\title{addTitle}
\usage{
add_title(
  out,
  title = NULL,
  size = 20,
  height = 30,
  color = "white",
  titleColor = "black",
  gravity = "west"
)
}
\arguments{
\item{out}{the image}

\item{title}{the title}

\item{size}{the size}

\item{height}{title height}

\item{color}{bg color}

\item{titleColor}{title color}

\item{gravity}{positioning of the title in the blank image}
}
\value{
output the image
}
\description{
Add the title to the image produced by output_overlay_image
using magick.
}
