% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_label_position}
\alias{append_label_position}
\title{append_label_position}
\usage{
append_label_position(g)
}
\arguments{
\item{g}{graph}
}
\value{
tbl_graph
}
\description{
Append the label position at center of edges
in global map like ko01100 where line type nodes
are present in KGML. 
Add `center` column to graph edge.
}
\examples{
## Simulate nodes containing `graphics_type` of line and `coords`
gm_test <- data.frame(name="ko:K00112",type="ortholog",reaction="rn:R00112",
           graphics_name="K00112",fgcolor="#ff0000",bgcolor="#ffffff",
           graphics_type="line",coords="1,2,3,4",orig.id=1,pathway_id="test")
gm_test <- tbl_graph(gm_test)
test <- process_line(gm_test) |> append_label_position()
}
