% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{carrow}
\alias{carrow}
\title{carrow}
\usage{
carrow(length = unit(2, "mm"))
}
\arguments{
\item{length}{arrow length in unit()}
}
\value{
arrow()
}
\description{
make closed type arrow
}
\examples{
carrow()
}
