% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_functions.R
\name{create_test_module}
\alias{create_test_module}
\title{create_test_module}
\usage{
create_test_module()
}
\value{
return a test module to use in examples
}
\description{
Test kegg_module for examples and vignettes.
The module has no biological meanings.
}
\examples{
create_test_module()
}
