% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{create_test_network}
\alias{create_test_network}
\title{create_test_network}
\usage{
create_test_network()
}
\value{
test network
}
\description{
create_test_network
}
\examples{
create_test_network()
}
