% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_parenthesis_pairs}
\alias{find_parenthesis_pairs}
\title{find_parenthesis_pairs
find pairs of parenthesis}
\usage{
find_parenthesis_pairs(s)
}
\description{
find_parenthesis_pairs
find pairs of parenthesis
}
