% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{geom_kegg}
\alias{geom_kegg}
\title{geom_kegg}
\usage{
geom_kegg(edge_color = NULL, node_label = name, group_color = "red")
}
\arguments{
\item{edge_color}{color attribute to edge}

\item{node_label}{column name for node label}

\item{group_color}{border color for group node rectangles}
}
\value{
ggplot2 object
}
\description{
convenient function for plotting KEGG pathway graph
add geom_node_rect, geom_node_text and geom_edge_link
}
\examples{
nodes <- data.frame(name=c("hsa:1029","hsa:4171"),
                   x=c(1,1),
                   xmin=c(-1,-1),
                   xmax=c(2,2),
                   y=c(1,1),
                   ymin=c(-1,-1),
                   ymax=c(2,2),
                   type=c("gene","gene"),
                   bgcolor=c("red","blue"))
edges <- data.frame(from=1, to=2)
graph <- tbl_graph(nodes, edges)
p <- ggraph(graph, layout="manual", x=x, y=y)+
geom_kegg()
}
