% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{geom_node_rect}
\alias{geom_node_rect}
\title{geom_node_rect}
\usage{
geom_node_rect(
  mapping = NULL,
  data = NULL,
  position = "identity",
  show.legend = NA,
  ...
)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data to plot}

\item{position}{positional argument}

\item{show.legend}{whether to show legend}

\item{...}{passed to `params` in `layer()` function}
}
\value{
geom
}
\description{
Plot rectangular shapes to ggplot2 using GeomRect,
using StatFilter in ggraph
}
\examples{
test_pathway <- create_test_pathway()
plt <- ggraph(test_pathway, layout="manual", x=x, y=y) +
 geom_node_rect()
}
