% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{geom_node_rect_multi}
\alias{geom_node_rect_multi}
\title{geom_node_rect_multi}
\usage{
geom_node_rect_multi(..., asIs = TRUE)
}
\arguments{
\item{...}{color columns}

\item{asIs}{treat the color as is or not}
}
\value{
ggplot2 object
}
\description{
Wrapper function for plotting multiple rects
with background color with geom_node_rect().
All columns should belong to the same scale when
using `asIs=FALSE`. If you need multiple scales for 
each element, please use `ggh4x::scale_fill_multi`
for each.
}
\examples{
plt <- create_test_pathway() \%>\% ggraph() + geom_node_rect_multi(bgcolor)
}
