% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{geom_node_shadowtext}
\alias{geom_node_shadowtext}
\title{geom_node_shadowtext}
\usage{
geom_node_shadowtext(
  mapping = NULL,
  data = NULL,
  position = "identity",
  show.legend = NA,
  ...
)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data to plot}

\item{position}{positional argument}

\item{show.legend}{whether to show legend}

\item{...}{passed to `params` in `layer()` function}
}
\value{
geom
}
\description{
Plot shadowtext at node position, use StatFilter in ggraph
}
\examples{
test_pathway <- create_test_pathway()
plt <- ggraph(test_pathway, layout="manual", x=x, y=y) +
 geom_node_shadowtext(aes(label=name))
}
