% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_functions.R
\name{get_module_attribute,kegg_module-method}
\alias{get_module_attribute,kegg_module-method}
\title{get_module_attribute}
\usage{
\S4method{get_module_attribute}{kegg_module}(x, attribute)
}
\arguments{
\item{x}{kegg_module class object}

\item{attribute}{slot name}
}
\value{
attribute of kegg_module
}
\description{
get the kegg_module class attribute
}
