% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_functions.R
\name{get_module_attribute}
\alias{get_module_attribute}
\title{get_module_attribute}
\usage{
get_module_attribute(x, attribute)
}
\arguments{
\item{x}{kegg_module class object}

\item{attribute}{pass to get_module_attribute}
}
\value{
attribute of kegg_module
}
\description{
Get slot from `kegg_module` class object.
}
\details{
Get slot from `kegg_module` class object.
}
