% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{get_network_attribute,kegg_network-method}
\alias{get_network_attribute,kegg_network-method}
\title{get_network_attribute}
\usage{
\S4method{get_network_attribute}{kegg_network}(x, attribute)
}
\arguments{
\item{x}{kegg_network class object}

\item{attribute}{slot name}
}
\value{
attribute of kegg_module
}
\description{
get the kegg_network class attribute
}
