% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{get_network_attribute}
\alias{get_network_attribute}
\title{get_network_attribute}
\usage{
get_network_attribute(x, attribute)
}
\arguments{
\item{x}{kegg_network class object}

\item{attribute}{pass to get_network_attribute}
}
\value{
attribute of kegg_network
}
\description{
get slot from `kegg_network` class
}
