% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkegg.R
\name{ggkegg}
\alias{ggkegg}
\title{ggkegg}
\usage{
ggkegg(
  pid,
  layout = "native",
  return_igraph = FALSE,
  convert_org = NULL,
  convert_first = TRUE,
  convert_collapse = NULL,
  convert_reaction = FALSE,
  delete_undefined = FALSE,
  delete_zero_degree = FALSE,
  numeric_attribute = NULL,
  node_rect_nudge = 0,
  group_rect_nudge = 2
)
}
\arguments{
\item{pid}{KEGG Pathway id e.g. hsa04110}

\item{layout}{default to "native", using KGML positions}

\item{return_igraph}{return the resulting igraph object}

\item{convert_org}{these organism names are fetched from REST API
and cached, and used to convert the KEGG identifiers.
e.g. c("hsa", "compound")}

\item{convert_first}{after converting, take the first element as 
node name when multiple genes are listed in the node}

\item{convert_collapse}{if not NULL, collapse the gene names by this character
when multiple genes are listed in the node.}

\item{convert_reaction}{reaction name (graph attribute `reaction`) 
will be converted to reaction formula}

\item{delete_undefined}{delete `undefined` node specifying group,
should be set to `TRUE` when the layout is not from native KGML.}

\item{delete_zero_degree}{delete nodes with zero degree, default to FALSE}
}
\description{
main function parsing KEGG pathway data,
making igraph object and passing it to ggraph.
}
