% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_functions.R
\name{ggkeggsave}
\alias{ggkeggsave}
\title{ggkeggsave}
\usage{
ggkeggsave(filename, plot, dpi = 300, wscale = 90, hscale = 90)
}
\arguments{
\item{filename}{file name of the image}

\item{plot}{plot to be saved}

\item{dpi}{dpi, passed to ggsave}

\item{wscale}{width scaling factor for pixel to inches}

\item{hscale}{height scaling factor fo pixel to inches}
}
\value{
save the image
}
\description{
save the image respecting the original width and height of the image.
Only applicable for the ggplot object including `overlay_raw_map` layers.
}
