% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplot_add.add_readable_edge_label}
\alias{ggplot_add.add_readable_edge_label}
\title{ggplot_add.geom_node_rect_kegg}
\usage{
\method{ggplot_add}{add_readable_edge_label}(object, plot, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{The ggplot object to add object to}

\item{object_name}{The name of the object to add}
}
\value{
ggplot2 object
}
\description{
ggplot_add.geom_node_rect_kegg
}
\examples{
test_pathway <- create_test_pathway()
plt <- ggraph(test_pathway, layout="manual", x=x, y=y) +
 geom_edge_link(aes(label=subtype_name, label_colour=subtype_name))+
 add_readable_edge_label()
}
