% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ggplot_add.geom_node_rect}
\alias{ggplot_add.geom_node_rect}
\title{ggplot_add.geom_node_rect}
\usage{
\method{ggplot_add}{geom_node_rect}(object, plot, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{The ggplot object to add object to}

\item{object_name}{The name of the object to add}
}
\description{
ggplot_add.geom_node_rect
}
