% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_functions.R
\name{ggplot_add.overlay_raw_map}
\alias{ggplot_add.overlay_raw_map}
\title{ggplot_add.overlay_raw_map}
\usage{
\method{ggplot_add}{overlay_raw_map}(object, plot, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{The ggplot object to add object to}

\item{object_name}{The name of the object to add}
}
\value{
ggplot2 object
}
\description{
ggplot_add.overlay_raw_map
}
\examples{
## Need `pathway_id` column in graph 
## if the function is to automatically infer
graph <- create_test_pathway() \%>\% mutate(pathway_id="hsa04110")
ggraph(graph) + overlay_raw_map()

}
