% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{ggplot_add.stamp}
\alias{ggplot_add.stamp}
\title{ggplot_add.stamp}
\usage{
\method{ggplot_add}{stamp}(object, plot, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{The ggplot object to add object to}

\item{object_name}{The name of the object to add}
}
\value{
ggplot2 object
}
\description{
ggplot_add.stamp
}
\examples{
test_pathway <- create_test_pathway()
plt <- ggraph(test_pathway, layout="manual", x=x, y=y) +
 stamp("hsa:6737")
}
