% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_functions.R
\name{highlight_entities}
\alias{highlight_entities}
\title{highlight_entities}
\usage{
highlight_entities(
  pathway,
  set,
  how = "any",
  name = "graphics_name",
  sep = ",",
  no_sep = FALSE,
  type = "gene",
  fill_color = "tomato",
  legend_name = NULL
)
}
\arguments{
\item{pathway}{pathway ID to be passed to `pathway()`}

\item{set}{vector of identifiers}

\item{how}{if `all`, if node contains multiple
IDs separated by `sep`, highlight if all the IDs
are in query. if `any`, highlight if one of the IDs
is in query.}

\item{name}{which column to search for}

\item{sep}{separater for node names}

\item{no_sep}{not separate node name}

\item{type}{entitie type, default to 'gene'}

\item{fill_color}{highlight color, default to 'tomato'}

\item{legend_name}{legend name, NULL to suppress}
}
\value{
overlaid map
}
\description{
highlight the entities in the pathway,
overlay raw map and return the results
}
\examples{
highlight_entities("hsa04110", c("CDKN2A"), legend_name="interesting")
}
