% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{highlight_module}
\alias{highlight_module}
\title{highlight_module}
\usage{
highlight_module(graph, kmo, name = "name", sep = " ")
}
\arguments{
\item{name}{which column to search for}

\item{sep}{separater for node names}

\item{set}{set of identifiers}
}
\description{
identify if edges are involved in module.
Using reaction of module.
}
