% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_functions.R
\name{highlight_module_compound}
\alias{highlight_module_compound}
\title{highlight_module_compound}
\usage{
highlight_module_compound(
  kmo,
  label,
  name = "name",
  sep = " ",
  verbose = FALSE
)
}
\arguments{
\item{kmo}{kegg_module class object which stores reaction}

\item{label}{edge label}

\item{name}{which column to search for}

\item{sep}{separator for node names}

\item{verbose}{show messages or not}
}
\description{
identify if nodes are involved in compounds in module reaction.
please note that all the compounds satisfying queried module's reaction
will be highlighted, thus it will not be exactly the same as KEGG mapper.
}
