% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_functions.R
\name{highlight_module_reaction}
\alias{highlight_module_reaction}
\title{highlight_module_reaction}
\usage{
highlight_module_reaction(kmo, name = "name", sep = " ", verbose = FALSE)
}
\arguments{
\item{kmo}{kegg_module class object which stores reaction}

\item{name}{which column to search for}

\item{sep}{separator for node names}

\item{verbose}{show messages or not}
}
\description{
identify if edges are involved in module reaction.
please note that all the reactions satisfying queried module's reaction
will be highlighted, thus it will not be exactly the same as KEGG mapper.
}
