% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_functions.R
\name{module}
\alias{module}
\title{module
KEGG module parsing function}
\usage{
module(mid, use_cache = FALSE, directory = NULL)
}
\arguments{
\item{mid}{KEGG module ID}
}
\value{
list of module definition and reaction
}
\description{
module
KEGG module parsing function
}
\examples{
\donttest{module("M00003")}
}
