% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_functions.R
\name{module_abundance}
\alias{module_abundance}
\title{module_abundance
weighted mean abundance of fraction of present KO in the block}
\usage{
module_abundance(mod_id, vec, num = 1, calc = "weighted_mean")
}
\arguments{
\item{mod_id}{module ID}

\item{vec}{KO-named vector of abundance without prefix `ko:`}

\item{num}{definition number when multiple definitions are present}

\item{calc}{calculation of final results, mean or weighted_mean}
}
\value{
numeric value
}
\description{
module_abundance
weighted mean abundance of fraction of present KO in the block
}
\examples{
\donttest{module_abundance("M00003",c(1.2) |> setNames("K00927"))}
}
