% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_functions.R
\name{module_completeness}
\alias{module_completeness}
\title{module_completeness}
\usage{
module_completeness(kmo, query, name = "1")
}
\arguments{
\item{kmo}{module object}

\item{query}{vector of KO}

\item{name}{name of definitions when multiple definitions are present}
}
\value{
tibble
}
\description{
module_completeness
}
\examples{
test_complete <- module_completeness(create_test_module(), c("K00112"))
}
