% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_functions.R
\name{module_text}
\alias{module_text}
\title{module_text
Obtain textual representation of modules for all the steps}
\usage{
module_text(kmo, candidate_ko = NULL, paint_colour = "tomato", convert = NULL)
}
\description{
module_text
Obtain textual representation of modules for all the steps
}
