% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{multi_pathway_native}
\alias{multi_pathway_native}
\title{multi_pathway_native}
\usage{
multi_pathway_native(pathways, row_num = 2, return_list = FALSE)
}
\arguments{
\item{pathways}{pathway vector}

\item{row_num}{row number}

\item{return_list}{return list of graphs instead of joined graph}
}
\value{
graph adjusted for the position
}
\description{
If you want to combine multiple KEGG pathways with their native coordinates,
supply this function a vector of pathway IDs and row number. This returns the
joined graph or list of graphs in which the coordinates are altered to panel
the pathways.
}
\examples{
## Pass multiple pathway IDs
multi_pathway_native(list("hsa04110","hsa03460"))

}
