% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{network}
\alias{network}
\title{KEGG network parsing function
network
parsing the network elements starting with N}
\usage{
network(nid)
}
\value{
list of network definition
}
\description{
KEGG network parsing function
network
parsing the network elements starting with N
}
