% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{network_graph}
\alias{network_graph}
\title{network_graph}
\usage{
network_graph(kne, type = "definition")
}
\arguments{
\item{kne}{network object}

\item{type}{definition or expanded}
}
\value{
tbl_graph
}
\description{
obtain tbl_graph of KEGG network
}
\examples{
ne <- create_test_network()
neg <- network_graph(ne)
}
