% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_functions.R
\name{obtain_sequential_module_definition}
\alias{obtain_sequential_module_definition}
\title{obtain_sequential_module_definition}
\usage{
obtain_sequential_module_definition(def, step = NULL)
}
\description{
Given module definition and step number,
Recursively obtain graphical represencation of step and 
connect them by pseudo-nodes representing steps.
}
