% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacyc_function.R
\name{parse_reaction_layout}
\alias{parse_reaction_layout}
\title{parse_reaction_layout}
\usage{
parse_reaction_layout(reacs)
}
\description{
parse reaction layout of metacyc pathway
and return tbl_graph object
}
