% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{pathway}
\alias{pathway}
\title{pathway}
\usage{
pathway(
  pid,
  group_rect_nudge = 2,
  node_rect_nudge = 0,
  invert_y = TRUE,
  add_pathway_id = TRUE,
  return_tbl_graph = TRUE,
  return_image = FALSE
)
}
\arguments{
\item{pid}{pathway id}

\item{group_rect_nudge}{nudge the position of group node
default to add slight increase to show the group node}

\item{node_rect_nudge}{nudge the position of all node}

\item{invert_y}{invert the y position to match with R graphics}

\item{add_pathway_id}{add pathway id to graph, default to TRUE
needed for the downstream analysis}

\item{return_tbl_graph}{return tbl_graph object, if FALSE, return igraph}

\item{return_image}{return the image URL}
}
\value{
tbl_graph by default
}
\description{
KEGG pathway parsing function
}
\examples{
\donttest{pathway("hsa04110")}
}
