% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_functions.R
\name{pathway_abundance}
\alias{pathway_abundance}
\title{pathway_abundance}
\usage{
pathway_abundance(id, vec, num = 1)
}
\arguments{
\item{id}{pathway id}

\item{vec}{named vector of abundance}

\item{num}{number of module definition}
}
\value{
numeric value
}
\description{
pathway_abundance
}
\examples{
\donttest{pathway_abundance("ko00270", c(1.2) |> `setNames`("K00927"))}
}
