% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{pathway_info}
\alias{pathway_info}
\title{pathway_info}
\usage{
pathway_info(pid, use_cache = FALSE, directory = NULL)
}
\arguments{
\item{pid}{KEGG Pathway id}

\item{use_cache}{whether to use cache}

\item{directory}{directory of file}
}
\value{
list of orthology and module contained in the pathway
}
\description{
obtain the list of pathway information
}
\examples{
pathway_info("hsa04110")
}
