% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_kegg_network}
\alias{plot_kegg_network}
\title{plot_kegg_network}
\usage{
plot_kegg_network(g)
}
\arguments{
\item{g}{graph object returned by `network()`}
}
\value{
ggplot2 object
}
\description{
plot the output of network_graph
}
\examples{
ne <- create_test_network()
neg <- network_graph(ne)
plt <- plot_kegg_network(neg)
}
