% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_module_blocks}
\alias{plot_module_blocks}
\title{plot_module_blocks}
\usage{
plot_module_blocks(all_steps, layout = "kk")
}
\arguments{
\item{all_steps}{the result of `obtain_sequential_module_definition()`}

\item{layout}{ggraph layout parameter}
}
\value{
ggplot2 object
}
\description{
wrapper function for plotting network representation of 
module definition blocks
}
\examples{
mo <- create_test_module()
## The output of `obtain_sequential_module_definition` 
## is used for `plot_module_blocks()`
sequential_mod <- obtain_sequential_module_definition(mo)
plt <- plot_module_blocks(sequential_mod)
}
