% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_module_steps}
\alias{plot_module_steps}
\title{plot_module_steps
wrapper function for plotting module definition steps}
\usage{
plot_module_steps(all_steps, layout = "kk")
}
\description{
plot_module_steps
wrapper function for plotting module definition steps
}
