% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_line_compounds}
\alias{return_line_compounds}
\title{return_line_compounds}
\usage{
return_line_compounds(g, orig)
}
\arguments{
\item{g}{tbl_graph object}

\item{orig}{original edge ID}
}
\value{
vector of original compound node IDs
}
\description{
In the map, where lines are converted to edges,
identify compounds that are linked by the reaction.
Give the original edge ID of KGML (orig.id in edge table), and 
return the original compound node ID
}
\examples{
## For those containing nodes with the graphic type of `line`
## This returns no IDs as no edges are present
gm_test <- create_test_pathway(line=TRUE)
test <- process_line(gm_test) \%>\% return_line_compounds(1)
}
