% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{return_pathway_example}
\alias{return_pathway_example}
\title{return_pathway_example}
\usage{
return_pathway_example()
}
\value{
tbl_graph
}
\description{
As downloading from KEGG API is not desirable
in vignettes or examples, return the `tbl_graph`
with two nodes and two edges.
}
\examples{
return_pathway_example
}
