% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp}
\alias{stamp}
\title{stamp}
\usage{
stamp(name, color = "red", which_column = "name")
}
\arguments{
\item{name}{name of the nodes}

\item{color}{color of the stamp}

\item{which_column}{which node column to search}
}
\value{
ggplot2 object
}
\description{
place stamp on the specified node
}
\examples{
test_pathway <- create_test_pathway()
plt <- ggraph(test_pathway, layout="manual", x=x, y=y) +
 stamp("hsa:6737")
}
