% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_with_bnlearn}
\alias{combine_with_bnlearn}
\title{combine_with_bnlearn}
\usage{
combine_with_bnlearn(pg, str, av, prefix = "ko:", how = "any")
}
\arguments{
\item{pg}{reference graph (output of `pathway`)}

\item{str}{strength data.frame}

\item{av}{averaged network to plot}

\item{prefix}{add prefix to node name of original averaged network
like, `hsa:` or `ko:`.}

\item{how}{`any` or `all`}
}
\value{
tbl_graph
}
\description{
combine the reference KEGG pathway graph 
with bnlearn boot.strength output
}
\examples{
if (requireNamespace("bnlearn", quietly = TRUE)) {
  av <- bnlearn::model2network("[1029|4171][4171]")
  str <- data.frame(from="4171",to="1029",strength=0.8,direction=0.7)
  nodes <- data.frame(name=c("hsa:1029","hsa:4171"),
                      x=c(1,1),
                      xmin=c(-1,-1),
                      xmax=c(2,2),
                      y=c(1,1),
                      ymin=c(-1,-1),
                      ymax=c(2,2))
  edges <- data.frame(from=1, to=2, name="K00112")
  graph <- tbl_graph(nodes, edges)
  combined <- combine_with_bnlearn(graph, str, av, prefix="hsa:")
}

}
