% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_id}
\alias{convert_id}
\title{convert_id}
\usage{
convert_id(
  org,
  name = "name",
  convert_column = NULL,
  colon = TRUE,
  first_arg_comma = TRUE,
  sep = " ",
  first_arg_sep = TRUE,
  divide_semicolon = TRUE,
  edge = FALSE
)
}
\arguments{
\item{org}{which identifier to convert}

\item{name}{which column to convert in edge or node table}

\item{convert_column}{which column is parsed in 
obtained data frame from KEGG REST API}

\item{colon}{whether the original ids include colon (e.g. `ko:`)
If `NULL`, automatically set according to `org`}

\item{first_arg_comma}{take first argument of comma-separated
string, otherwise fetch all strings}

\item{sep}{separater to separate node names, defaul to space}

\item{first_arg_sep}{take first argument if multiple identifiers
are in the node name, otherwise parse all identifiers}

\item{divide_semicolon}{whether to divide string by semicolon,
and take the first value}

\item{edge}{if converting edges}
}
\value{
vector containing converted IDs
}
\description{
convert the identifier using retrieved information
}
\examples{
nodes <- data.frame(name=c("hsa:1029","hsa:4171"),
                   x=c(1,1),
                   xmin=c(-1,-1),
                   xmax=c(2,2),
                   y=c(1,1),
                   ymin=c(-1,-1),
                   ymax=c(2,2))
edges <- data.frame(from=1, to=2)
graph <- tbl_graph(nodes, edges)
\donttest{graph <- graph |> mutate(conv=convert_id("hsa"))}

}
