% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edge_matrix}
\alias{edge_matrix}
\title{edge_matrix}
\usage{
edge_matrix(
  graph,
  mat,
  gene_type = "SYMBOL",
  org = "hsa",
  org_db = org.Hs.eg.db,
  num_combine = mean
)
}
\arguments{
\item{graph}{tbl_graph to append values to}

\item{mat}{matrix representing gene as row and sample as column}

\item{gene_type}{gene ID of matrix row}

\item{org}{organism ID to convert ID}

\item{org_db}{organism database to convert ID}

\item{num_combine}{function to combine multiple numeric values}
}
\value{
tbl_graph
}
\description{
given the matrix representing gene as row and sample as column,
append the edge value (sum of values of connecting nodes) to edge matrix and
return tbl_graph object
}
\examples{
nodes <- data.frame(name=c("hsa:1029","hsa:4171"),
                   x=c(1,1),
                   xmin=c(-1,-1),
                   xmax=c(2,2),
                   y=c(1,1),
                   ymin=c(-1,-1),
                   ymax=c(2,2))
edges <- data.frame(from=1, to=2, name="K00112")
graph <- tbl_graph(nodes, edges)
num_df <- data.frame(row.names=c("1029","4171"),
                     "sample1"=c(1.1,1.2),
                     "sample2"=c(1.1,1.2),
                     check.names=FALSE)
graph <- graph |> edge_matrix(num_df, gene_type = "ENTREZID")
}
