% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{geom_node_rect}
\alias{geom_node_rect}
\title{geom_node_rect}
\usage{
geom_node_rect(
  mapping = NULL,
  data = NULL,
  position = "identity",
  show.legend = NA,
  ...
)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data to plot}

\item{position}{positional argument}

\item{show.legend}{whether to show legend}

\item{...}{passed to `params` in `layer()` function}
}
\value{
geom
}
\description{
add rectangular shapes to ggplot2 using GeomRect,
using StatFilter in ggraph
}
\examples{
nodes <- data.frame(name=c("hsa:1029","hsa:4171"),
                   x=c(1,1),
                   xmin=c(-1,-1),
                   xmax=c(2,2),
                   y=c(1,1),
                   ymin=c(-1,-1),
                   ymax=c(2,2))
edges <- data.frame(from=1, to=2)
graph <- tbl_graph(nodes, edges)
plt <- ggraph(graph, layout="manual", x=x, y=y) +
 geom_node_rect()
}
