% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{geom_node_rect}
\alias{geom_node_rect}
\title{geom_node_rect}
\usage{
geom_node_rect(with_group = TRUE)
}
\arguments{
\item{with_group}{grouping rectangle will be drawn}

\item{g}{ggraph object}
}
\description{
geom_node_rect
}
