% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_functions.R
\name{ggkeggsave}
\alias{ggkeggsave}
\title{ggkeggsave}
\usage{
ggkeggsave(filename, plot, dpi = 300)
}
\arguments{
\item{filename}{file name of the image}

\item{plot}{plot to be saved}

\item{dpi}{dpi, passed to ggsave}
}
\value{
save the image
}
\description{
ggkeggsave
}
