% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ggplot_add.geom_node_rect_kegg}
\alias{ggplot_add.geom_node_rect_kegg}
\title{ggplot_add.geom_node_rect_kegg}
\usage{
\method{ggplot_add}{geom_node_rect_kegg}(object, plot, ...)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{The ggplot object to add object to}

\item{...}{The other arguments}
}
\value{
ggplot2 object
}
\description{
ggplot_add.geom_node_rect_kegg
}
\examples{
test_pathway <- create_test_pathway()
plt <- ggraph(test_pathway, layout="manual", x=x, y=y) +
 geom_node_rect_kegg(type="gene")
}
