% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_functions.R
\name{ggplot_add.overlay_raw_map}
\alias{ggplot_add.overlay_raw_map}
\title{ggplot_add.overlay_raw_map}
\usage{
\method{ggplot_add}{overlay_raw_map}(object, plot, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{The ggplot object to add object to}

\item{object_name}{The name of the object to add}
}
\value{
ggplot2 object
}
\description{
ggplot_add.overlay_raw_map
}
\examples{
nodes <- data.frame(name=c("hsa:1029","hsa:4171"),
                   x=c(1,1),
                   xmin=c(-1,-1),
                   xmax=c(2,2),
                   y=c(1,1),
                   ymin=c(-1,-1),
                   ymax=c(2,2),
                   pathway_id="hsa04110")
edges <- data.frame(from=1, to=2, name="K00112")
graph <- tbl_graph(nodes, edges)
\dontrun{ggraph(graph) + overlay_raw_map()}

}
