% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_functions.R
\name{highlight_module}
\alias{highlight_module}
\title{highlight_module}
\usage{
highlight_module(graph, kmo, name = "name", sep = " ")
}
\arguments{
\item{graph}{graph}

\item{kmo}{kegg_module class object}

\item{name}{which column to search for}

\item{sep}{separater for node names}
}
\description{
Using reaction and compound IDs in the module,
highlight the interactions involved.
Please see `highlight_module_reaction` and `highlight_module_compound`.
}
