% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{highlight_set_edges}
\alias{highlight_set_edges}
\title{highlight_set_edges}
\usage{
highlight_set_edges(set, name = "name", sep = " ")
}
\arguments{
\item{set}{set of identifiers}

\item{name}{which column to search for}

\item{sep}{separater for node names}
}
\description{
identify if edges are involved in specific queries.
}
