% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_functions.R
\name{highlight_set_nodes}
\alias{highlight_set_nodes}
\title{highlight_set_nodes}
\usage{
highlight_set_nodes(set, name = "name", sep = " ")
}
\arguments{
\item{set}{set of identifiers}

\item{name}{which column to search for}

\item{sep}{separater for node names}
}
\description{
identify if nodes are involved in specific queriy.
if multiple IDs are listed after separation by `sep`,
only return TRUE if all the IDs are in the query.
}
