% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_functions.R
\name{module_text}
\alias{module_text}
\title{module_text
Obtain textual representation of module definition for all the blocks}
\usage{
module_text(
  kmo,
  name = "1",
  candidate_ko = NULL,
  paint_colour = "tomato",
  convert = NULL
)
}
\description{
module_text
Obtain textual representation of module definition for all the blocks
}
