% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_functions.R
\name{module_text}
\alias{module_text}
\title{module_text
Obtain textual representation of module definition for all the blocks}
\usage{
module_text(
  kmo,
  name = "1",
  candidate_ko = NULL,
  paint_colour = "tomato",
  convert = NULL
)
}
\arguments{
\item{kmo}{module object}

\item{name}{name of definition}

\item{candidate_ko}{KO to highlight}

\item{paint_colour}{color to highlight}

\item{convert}{named vector converting the KO to gene name}
}
\value{
textual description of module definitions
}
\description{
module_text
Obtain textual representation of module definition for all the blocks
}
