% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{network}
\alias{network}
\title{KEGG network parsing function
network
parsing the network elements starting with N}
\usage{
network(nid, use_cache = FALSE, directory = NULL)
}
\arguments{
\item{nid}{KEGG NETWORK ID}

\item{use_cache}{use cache}

\item{directory}{directory to save raw files}
}
\value{
list of network definition
}
\description{
KEGG network parsing function
network
parsing the network elements starting with N
}
\examples{
\dontrun{network("N00002")}
}
