% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_functions.R
\name{obtain_sequential_module_definition}
\alias{obtain_sequential_module_definition}
\title{obtain_sequential_module_definition}
\usage{
obtain_sequential_module_definition(kmo, block = NULL)
}
\description{
Given module definition and block number,
Recursively obtain graphical represencation of block and 
connect them by pseudo-nodes representing blocks.
}
