% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{pathway}
\alias{pathway}
\title{pathway}
\usage{
pathway(
  pid,
  rect_width = 25,
  rect_height = 10,
  convert_org = NULL,
  convert_collapse = NULL,
  convert_first = TRUE,
  group_rect_nudge = 2,
  node_rect_nudge = 0,
  invert_y = TRUE,
  add_pathway_id = TRUE,
  return_tbl_graph = TRUE,
  return_image = FALSE
)
}
\arguments{
\item{rect_width}{manual specification of rectangle}

\item{rect_height}{manual specification of rectangle}

\item{convert_org}{organism or category for converting the name
e.g. "ko", "hsa", ...}

\item{add_pathway_id}{add pathway id to graph}

\item{return_tbl_graph}{return tbl_graph object}
}
\value{
tbl_graph by default
}
\description{
KEGG pathway parsing function
}
