% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_module_text}
\alias{plot_module_text}
\title{plot_module_text
plot the text representation of KEGG modules}
\usage{
plot_module_text(plot_list, show_name = "name")
}
\description{
plot_module_text
plot the text representation of KEGG modules
}
