% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{process_line}
\alias{process_line}
\title{process_line}
\usage{
process_line(g, invert_y = TRUE, verbose = FALSE)
}
\arguments{
\item{g}{graph}

\item{invert_y}{whether to invert the position, default to TRUE
should match with `pathway` function}

\item{verbose}{show progress}
}
\value{
tbl_graph
}
\description{
process the KGML containing graphics type of `line`, like
global maps e.g. ko01100. Recursively add nodes and edges 
connecting them based on `coords` properties in KGML.
}
\details{
We cannot show directed arrows, as coords are not ordered to show direction.
}
\examples{
## For those containing nodes with the graphic type of `line`,
## parse the coords attributes to edges.
gm_test <- create_test_pathway(line=TRUE)
test <- process_line(gm_test)
}
