% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{process_line}
\alias{process_line}
\title{process_line}
\usage{
process_line(g, invert_y = TRUE, verbose = FALSE)
}
\arguments{
\item{g}{graph}

\item{invert_y}{whether to invert the position, default to TRUE
should match with `pathway` function}

\item{verbose}{show progress}
}
\value{
tbl_graph
}
\description{
process the KGML containing graphics type of `line`, like
global maps e.g. ko01100. Recursively add nodes and edges 
connecting them based on `coords` properties in KGML.
}
\examples{
## For those containing nodes with the graphic type of `line`
gm_test <- data.frame(name="ko:K00112",type="ortholog",reaction="rn:R00112",
           graphics_name="K00112",fgcolor="#ff0000",bgcolor="#ffffff",
           graphics_type="line",coords="1,2,3,4",orig.id=1,pathway_id="test")
gm_test <- tbl_graph(gm_test)
test <- process_line(gm_test)
}
