% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{process_line}
\alias{process_line}
\title{process_line
process the kgml containing graphics type of `line`
e.g. in ko01100}
\usage{
process_line(g, invert_y = TRUE, verbose = FALSE)
}
\arguments{
\item{g}{graph}

\item{invert_y}{whether to invert the position, default to TRUE
should match with `pathway` function}

\item{verbose}{show progress}
}
\value{
tbl_graph
}
\description{
process_line
process the kgml containing graphics type of `line`
e.g. in ko01100
}
