% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{process_reaction}
\alias{process_reaction}
\title{process_reaction}
\usage{
process_reaction(g, single_edge = FALSE)
}
\arguments{
\item{g}{graph}

\item{single_edge}{discard one edge when edge type is `reversible`}
}
\value{
tbl_graph
}
\description{
process the kgml of global maps
e.g. in ko01100
}
\details{
Typically, `process_line` function is used to draw relationships 
as in the original KGML positions, however, the `coords` properties
is not considering the direction of reactions (substrate -> product),
thus if it is preferred, `process_reaction` is used to populate
new edges corresponding to `substrate -> product` and `product -> substrate`
if the reaction is reversible.
}
\examples{
gm_test <- rbind(data.frame(name="cpd:C99998",type="compound",
           graphics_name="C99998",fgcolor="#ff0000",bgcolor="#ffffff"),
           data.frame(name="cpd:C99999",type="compound",
                      graphics_name="C99999",fgcolor="#ff0000",bgcolor="#ffffff"),
           data.frame(name="ko:K00224",type="ortholog",
                      graphics_name="K00224",fgcolor="#ff0000",bgcolor="#ffffff")
           )
gm_test_edges <- rbind(data.frame(from=1,to=3,reaction="rn:R99999",subtype_name="substrate",
                            type="irreversible"),
                       data.frame(from=3,to=2,reaction="rn:R99999",subtype_name="product",
                                  type="irreversible"))
gm_test <- tbl_graph(gm_test, gm_test_edges)
test <- process_reaction(gm_test)

}
